// types/destination.ts
export interface Destination {
  id: string
  name: string
  location: string
  description: string
  longDescription: string
  activities: string
  imageUrl: string
  rating: number
  reviewCount: number
  latitude: number
  longitude: number
  bestTimeToVisit: string
  recommendedDuration: string
  category: string[]
  region: string
}